feat: [AD-008] split view into smaller components
