+++ view/src/components/AnomalyList.jsx
@@ -1,0 +1,12 @@
+export default function AnomalyList({ filteredFeatures }) {
+  return (
+    <div>
+      {filteredFeatures.map((f, i) => (
+        <div key={i} style={{ marginBottom: 6 }}>
+          <strong>{f.properties.anomaly_id}</strong> <br />
+          {f.properties.timestamp}
+        </div>
+      ))}
+    </div>
+  );
+}
